+++ src/App.js
@@ -2,13 +2,13 @@ import React, { Component } from 'react';
 import './App.css';
 
 
-import Login from './components/Login';
+import SignIn from './components/auth/SignIn';
 class App extends Component {
   render() {
     return (
       <div className="container App">
         <header className="App-header">
-          <Login></Login>
+          <SignIn></SignIn>
         </header>
       </div>
     );
